Address Herman's review comment
